Keep l2norm k in fp32

+++ fla/ops/gdn2/chunk.py
@@ -99,11 +99,11 @@ def chunk_gdn2(
         scale = k.shape[-1] ** -0.5
 
     if use_qk_l2norm_in_kernel:
-        q, k = l2norm(q), l2norm(k)
+        q, k = l2norm(q), l2norm(k, output_dtype=torch.float)
     kb = beta.unsqueeze(-1) * k
     a = k * g.float().exp()
     b = -kb
-    k = kb.to(k.dtype)
+    k = kb.to(q.dtype)
 
     o, final_state = chunk_dplr_delta_rule(
         q=q,
